Fix footnotes (#230)
* Remove unreferenced Footnotes

* Insert Link to new chapter

Apparently the #TODO is now DONE ;-)

* Use consecutive numbering for footnotes

Some component seems to treat footnotes wrongly if there are numbers
skipped. In the text the numbers are used as written in the file, in the
footnote section the numbers are ascending. This leads to somewhat
strange behavior as references 1 and 3 in the text reference to
footnote 1 and 2 respectively